Added Pupify to the Portfolio
Probably the least annoying thing in this portfolio update. Both the HL2 level and the Audio Visuallizer need to be updated at their source and on the portfolio. HL2 could use minimal lighting and enemy updates then it's entire section in needed on the portfolio. I simply need to change the cloud service that the audio visualizer runs and builds on because it is way too slow currently, then the updates need to be explained in the portfolio